[ADD] l10n_de_cancellation_invoice
This commit adds a must have module for the German localization: Invoice cancellation.

Germany has specific requirements for credit notes:

  - A credit note must have the wording "Stornorechnung" or "Korrekturrechnung", the wordning "Gutschrift"
    as it's with standard Odoo is NOT allowed
  - The amounts must be printed signed (printed with a "-" sign) in the invoice document
  - The original invoice and it's date, what is credited, must be referred

This module makes Odoo compliant with those requirements. It computes the credit type
(Stornorechnung / Korrekturrechnung) and exchanges the standard Odoo term "Gutschrift" with the correct
one. Also the monetary amounts in the invoice document are signed.